+++ tap_harvest/schema.py
@@ -75,7 +75,6 @@ def get_schemas() -> Tuple[Dict, Dict]:
 
     return schemas, field_metadata
 
-
 def write_schema(stream, client, streams_to_sync, catalog) -> None:
     """Collect nested child streams to sync and write schema for selected
     streams."""